perf tools: Record sampling time for each entry
Sometimes users might want see a relation between value and time of a
sample.  And the time itself is a very important information to have.
As the sample->period value is a diff of counter values between samples
it might make sense to add time-diff between samples into a hist entry.

To do that, calculate elapsed time for each sample and record it.  The
elapsed time is the diff between current sample->time and previous
sample->time which was saved for each evsel.

Maybe we can use PERF_SAMPLE_READ for the precise result.

Reviewed-by: Pekka Enberg <penberg@kernel.org>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>