perf tools: Insert filtered entries to hists also
Currently if a sample was filtered by command line option, it just
dropped.  But this affects final output in that the percentage can be
different since the filtered entries were not included to the total.

For example, if an original output looked like below:

  $ perf report --stdio -s comm

  # Overhead  Command
  # ........  .......
  #
      74.19%      cc1
       7.61%      gcc
       6.11%       as
       4.35%       sh
       4.14%     make
       1.13%   fixdep

If we gave a filter, the output changed like below:

  $ perf report --stdio -s comm -c gcc,as

  # Overhead  Command
  # ........  .......
  #
      55.49%      gcc
      44.51%       as

But user might want to see this:

  $ perf report --stdio -s comm -c gcc,as

  # Overhead  Command
  # ........  .......
  #
       7.61%      gcc
       6.11%       as

Signed-off-by: Namhyung Kim <namhyung@kernel.org>