perf callchain: Count self callchain hits too
When perf report displays callchains, it only counts children hit
counts so even if some callchain node has self hits it wasn't
displayed in the output.  It can be easily see when using -g caller
option.

For example, in the below case 'c' was self-sampled most of time, but
in a very rare case (53.07 - 52.99 = 0.08%) it was sampled during
handling a timer interrupt.  But the output seems as if it always
calls the interrupt handler.

  $ perf report -g caller
    Children      Self  Command  Shared Object      Symbols
  ---------------------------------------------------------------

  -   53.07%    52.99%  abc2     abc2               [.] c
     - c
        + 100.00% apic_timer_interrupt

With this patch applied, it now shows like following:

  -   53.07%    52.99%  abc2     abc2               [.] c
     - c
        + 0.15% apic_timer_interrupt

Cc: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>