perf trace: Create struct thread for command line workload
When perf creates a new child to profile, the events are enabled on
exec().  And in this case, it doesn't synthesize any event for the
child since they'll be generated during exec().  But there's an window
between the enabling and the event generation.  This leads to some
early event not having a comm but pid (like ':15328').

Fix it by using perf_evlist__start_workload_ex() to create
corresponding thread at the same time.

Acked-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>